v0.2
faster filter-filter training
let attention influence attention passing

+++ dataset/question_dataset.py
@@ -62,8 +62,8 @@ class Dataset(torch.utils.data.Dataset):
             info.visual_dataset =\
                 teddy_dataset.ToyVisualDataset(args, info)
             all_indexes = np.arange(args.size_toy)
-            split_train = args.size_toy * 0.7
-            split_val = args.size_toy * 0.9
+            split_train = int(args.size_toy * 0.7)
+            split_val = int(args.size_toy * 0.9)
             cls.split_indexes = {'train': all_indexes[:split_train],
                                  'val': all_indexes[split_train: split_val],
                                  'test': all_indexes[split_val:]}
